update with the floatting number for push template email on experience year

+++ internal/handlers/presentation_email.go
@@ -15,24 +15,24 @@ func GeneratePresentationEmailV2(c *gin.Context) {
 	// Structure intermédiaire pour gérer les langues flexibles
 	var tempReq struct {
 		CandidateData struct {
-			Prenom                    string `json:"prenom" binding:"required"`
-			TitrePoste                string `json:"titre_poste" binding:"required"`
-			NombreExperience          int    `json:"nombre_experience" binding:"required"`
-			Disponibilite             string `json:"disponibilite" binding:"required"`
-			Mobilite                  string `json:"mobilite" binding:"required"`
-			Diplome                   string `json:"diplome" binding:"required"`
-			Langues                   string `json:"langues"`
-			Certifications            string `json:"certifications"`
-			Hobbies                   string `json:"hobbies"`
-			Experience                string `json:"experience"`
-			ExperienceCount           int    `json:"experience_count"`
-			Logiciel                  string `json:"logiciel"`
-			Logiciels                 string `json:"logiciels"`
-			LogicielsCount            int    `json:"logiciels_count"`
-			CompetencesTechniques     string `json:"competences_techniques"`
-			CompetencesFonctionnelles string `json:"competences_fonctionnelles"`
-			Projets                   string `json:"projets"`
-			ProjetsCount              int    `json:"projets_count"`
+			Prenom                    string  `json:"prenom" binding:"required"`
+			TitrePoste                string  `json:"titre_poste" binding:"required"`
+			NombreExperience          float64 `json:"nombre_experience" binding:"required"`
+			Disponibilite             string  `json:"disponibilite" binding:"required"`
+			Mobilite                  string  `json:"mobilite" binding:"required"`
+			Diplome                   string  `json:"diplome" binding:"required"`
+			Langues                   string  `json:"langues"`
+			Certifications            string  `json:"certifications"`
+			Hobbies                   string  `json:"hobbies"`
+			Experience                string  `json:"experience"`
+			ExperienceCount           int     `json:"experience_count"`
+			Logiciel                  string  `json:"logiciel"`
+			Logiciels                 string  `json:"logiciels"`
+			LogicielsCount            int     `json:"logiciels_count"`
+			CompetencesTechniques     string  `json:"competences_techniques"`
+			CompetencesFonctionnelles string  `json:"competences_fonctionnelles"`
+			Projets                   string  `json:"projets"`
+			ProjetsCount              int     `json:"projets_count"`
 		} `json:"candidateData" binding:"required"`
 		Need       *string                      `json:"need,omitempty"`
 		TemplateID *string                      `json:"templateId,omitempty"`
